- Fixed comments in issue #10450: Multiple cronjob parts run simultaneously.   ( Merged from stable/3.8 rev. 18779, stable/3.9 rev. 18780 )
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/trunk@18782 a01eee8c-daba-0310-acae-fa49f3023285

+++ lib/ezutils/classes/ezmutex.php
@@ -244,7 +244,9 @@ class eZMutex
         else
         {
             $stealPid = $stealMutex->meta( 'pid' );
-            if ( function_exists( 'posix_kill' ) )
+            if ( is_numeric( $stealPid ) &&
+                 $stealPid != 0 &&
+                 function_exists( 'posix_kill' ) )
             {
                 eZDebug::writeNotice( 'Killing steal mutex process: ' . $stealPid );
                 posix_kill( $stealPid, 9 );
